+++ ansible_collections/arista/avd/plugins/action/yaml_templates_to_facts.py
@@ -8,9 +8,9 @@ from collections import ChainMap
 from ansible.plugins.action import ActionBase
 from ansible.errors import AnsibleActionFail
 from ansible.utils.vars import isidentifier
-from ansible_collections.arista.avd.plugins.module_utils.strip_empties import strip_null_from_data
+from ansible_collections.arista.avd.plugins.plugin_utils.strip_empties import strip_null_from_data
 from ansible_collections.arista.avd.plugins.plugin_utils.merge import merge
-from ansible_collections.arista.avd.plugins.module_utils.utils import template as templater, compile_searchpath, get
+from ansible_collections.arista.avd.plugins.plugin_utils.utils import template as templater, compile_searchpath, get
 from datetime import datetime
