+++ ansible_collections/arista/avd/plugins/plugin_utils/schema/avdvalidator.py
@@ -4,7 +4,7 @@ __metaclass__ = type
 import os
 import json
 import copy
-from ansible_collections.arista.avd.plugins.module_utils.utils import get_all
+from ansible_collections.arista.avd.plugins.plugin_utils.utils import get_all
 
 try:
     import jsonschema.validators
