+++ ansible_collections/arista/avd/plugins/plugin_utils/strip_empties.py
